Refactor time limit components
